chore: use save contex for json editor
